// src/app/pages/apple-watch/enum/apple-watch.enum.ts
export enum SizeMapTitle {
  forty = '40 mm',
  fortyOne = '41 mm',
  fortyFour = '44 mm',
  fortyFive = '45 mm',
}
export enum SizeMap {
  forty = 1,
  fortyOne = 2,
  fortyFour = 3,
  fortyFive = 4,
}
export const ParsSizesToString: any = {
  [SizeMap.forty]: '40 mm',
  [SizeMap.fortyOne]: '41 mm',
  [SizeMap.fortyFour]: '44 mm',
  [SizeMap.fortyFive]: '45 mm',
};